+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## To be Released
 
+* resource(scalingo_project): create the `project` schema with all methods
+* data_provider(scalingo_project): add the `project` source
+
 # 2.4.2
 
 * fix(ci): migrate goreleaser to v2